pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telemetry-platform"
version = "0.1.0"
description = "Telemetry ingestion + alerting platform (FastAPI + Postgres + Redis + Celery)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "redis>=7.1.0",
  "celery>=5.3",
  "httpx>=0.27.0",
  "slowapi>=0.1.1",
  "bcrypt>=4.0.1",
  "structlog>=22.2.0",
]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=html --cov-report=term-missing"

[tool.setuptools]
packages = ["app"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "pytest-mock>=3.12",
  "httpx>=0.27.0",
  "ruff>=0.1.15",
  "black>=24.1",
  "mypy>=1.8",
  "pre-commit>=3.6",
  "testcontainers[postgres]>=3.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false